requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.7
cssselect==1.3.0
Django==5.2.6
fake-useragent==2.2.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-ollama==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.4.25
lxml==6.0.1
lxml_html_clean==0.4.2
ollama==0.5.3
orjson==3.11.3
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.55.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyppeteer==2.0.0
pyppeteer_stealth==2.7.4
pyquery==2.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-html==0.10.0
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.23.0
zstandard==0.24.0
